+++ backend/src/main/java/com/devsimple/springmvc/domain/model/Cidade.java
@@ -12,7 +12,7 @@ public class Cidade implements Serializable {
     @Id
     @EqualsAndHashCode.Include
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String nome;
 
     @ManyToOne
@@ -22,18 +22,19 @@ public class Cidade implements Serializable {
     public Cidade() {
     }
 
-    public Cidade(Long id, String nome, Estado estado) {
+    public Cidade(Integer id, String nome, Estado estado) {
         super();
         this.id = id;
         this.nome = nome;
         this.estado = estado;
     }
 
-    public Long getId() {
+
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
